menambahkan cors dan membuat endpoint untuk memparsing token

+++ backend/app/controllers/lessons.controller.js
@@ -106,7 +106,7 @@ class LessonsController extends Controller {
             await models.lessons.destroy({
                 where: { name },
             })
-            ;("../../public/images/lessons")
+                ; ("../../public/images/lessons")
             removeFile(
                 path.join(
                     __dirname,
@@ -131,7 +131,9 @@ class LessonsController extends Controller {
         }
         const errors = validationResult(request)
         if (!errors.isEmpty()) {
-            removeFile(request.file.path)
+            if (request.file) {
+                removeFile(request.file.path)
+            }
             return this.error(errors)
         }
         //  console.log(request.file)
